Added tests for calculating super

+++ PayslipProblemVer2/Employee.cs
@@ -5,6 +5,6 @@ namespace PayslipProblemVer2
         public string Name { get; set; }
         public string Surname { get; set; }
         public uint AnnualSalary { get; set; }
-        public int SuperRate { get; set; }
+        public double SuperRate { get; set; }
     }
 }
